Move posts components to articles and fetch posts from API
