+++ app/articles/_containers/SectionPosts/containers/PostCard/index.module.scss
@@ -109,6 +109,7 @@
 
 .padding_2{
   padding-left: 16px;
+  padding-right: 16px;
   padding-top: 16px;
   margin-bottom: 16px;
 }
